lib: Move periodic to lib
fixes #1960
